Rename xvfb_init to xvfb_init.sh (#13)
Reason: the line endings in this file should be `lf`, and `git` can
convert them to `crlf` when checking out on Windows. The script with
incorrect line endings won't run in Linux container properly if an
attempt to launch devcontainer is done on Windows with Docker.

Specifying the extension will help with handling line endings when
checking out the repository on Windows.